style: Step01~05 파일을 폴더로 분리하여 이동
