--- src/components/onboarding/Step02.tsx
@@ -1,11 +1,0 @@
-import * as S from '../common/title/Title.style';
-
-const ThumbnailInput = () => {
-  return (
-    <S.TitleWrapper>
-      <S.Title>썸네일을 등록해주세요</S.Title>
-    </S.TitleWrapper>
-  );
-};
-
-export default ThumbnailInput;
