--- src/components/onboarding/Step03.tsx
@@ -1,13 +1,0 @@
-import * as S from '../common/title/Title.style';
-
-const GiftDelivery = () => {
-  return (
-    <S.TitleWrapper>
-      <S.Title>님께</S.Title>
-      <S.Title>언제 선물을</S.Title>
-      <S.Title>전달하실 예정인가요?</S.Title>
-    </S.TitleWrapper>
-  );
-};
-
-export default GiftDelivery;
