--- src/components/onboarding/Step04.tsx
@@ -1,13 +1,0 @@
-import * as S from '../common/title/Title.style';
-
-const SetTournamentSchedule = () => {
-  return (
-    <S.TitleWrapper>
-      <S.Title>선물 토너먼트</S.Title>
-      <S.Title>시작 일정을 설정해주세요</S.Title>
-      <S.SubTitle>토너먼트 시작 전까지 선물을 등록할 수 있어요.</S.SubTitle>
-    </S.TitleWrapper>
-  );
-};
-
-export default SetTournamentSchedule;
