--- src/components/onboarding/Step05.tsx
@@ -1,13 +1,0 @@
-import * as S from '../common/title/Title.style';
-
-const SetTournamentDuration = () => {
-  return (
-    <S.TitleWrapper>
-      <S.Title>선물 토너먼트</S.Title>
-      <S.Title>진행 시간을 설정해주세요</S.Title>
-      <S.SubTitle>토너먼트가 아래 시간 동안 진행돼요.</S.SubTitle>
-    </S.TitleWrapper>
-  );
-};
-
-export default SetTournamentDuration;
