+++ src/components/onboarding/step02/Step02.tsx
@@ -1,0 +1,12 @@
+import Title from '../../common/title/Title';
+import * as S from '../../common/title/Title.style';
+
+const ThumbnailInput = () => {
+  return (
+    <S.TitleWrapper>
+      <Title title='썸네일을 등록해주세요' />
+    </S.TitleWrapper>
+  );
+};
+
+export default ThumbnailInput;
