+++ src/components/onboarding/step03/Step03.tsx
@@ -1,0 +1,14 @@
+import Title from '../../common/title/Title';
+import * as S from '../../common/title/Title.style';
+
+const GiftDelivery = () => {
+  return (
+    <S.TitleWrapper>
+      <Title title='님께' />
+      <Title title='언제 선물을' />
+      <Title title='전달하실 예정인가요?' />
+    </S.TitleWrapper>
+  );
+};
+
+export default GiftDelivery;
